fix: 새해 페이지 url 변경
